requirements.txt
pymupdf4llm>=0.0.27
pytest>=7.0.0
pytest-cov>=4.0.0
jupyter>=1.0.0
notebook>=6.4.0
pandas>=1.5.0
numpy>=1.21.0
PyMuPDF>=1.21.0
Pillow>=9.0.0
pytesseract>=0.3.10

# Visual PDF Extraction dependencies
pdf2image>=1.17.0,<2
pdfplumber>=0.9.0
opencv-python>=4.8.0

# Visual Element Extraction dependencies
scikit-image>=0.25.0
matplotlib>=3.10.0
plotly>=6.3.0

# Schema validation
jsonschema>=4.17.0

# Retrieval demo (TF-IDF)
scikit-learn>=1.3.0

# Embeddings (CPU-friendly)
fastembed>=0.3.5

# Database and env
psycopg[binary,pool]>=3.2.0
python-dotenv>=1.0.1

# API
fastapi>=0.112.0
uvicorn>=0.30.0
qdrant-client>=1.10.0
